requirements.txt
flask==3.1.0
pillow==11.2.0
torch==2.7.0
torchvision==0.22.0
